requirements.txt
streamlit==1.45.0
ollama==0.6.1
google-genai==1.12.1
python-dotenv==1.1.0
reportlab==4.4.1